requirements.txt
alt-profanity-check # experimentalion of giardrail
google-generativeai
guardrails-ai
python-dotenv